rename: 컴포넌트 파일명을 컨벤션에 맞게 변경 (#15)
